# FIX: Further optimize mentioning groups in chat messages (part 2)